Multi_delete_user, delete_user, permission delete

+++ php_assets/delete_user.php
@@ -1,10 +1,20 @@
-<?php session_start();
-    require_once("connectdb.php");
-    require ("fonctions.php");
+<?php
 
-$pdoDe = $conn->prepare('DELETE FROM utilisateurs WHERE id= ? LIMIT 1');
-$pdoDe->execute([
-nettoyage($_GET['id'])
-]);
+require_once("connectdb.php");
+require("verif_session_connect.php");
+require("fonctions.php");
+require("permission.php");
+
+if(isset($_GET['id']) && $_GET['id'] != ""){
+
+    // On vérifie si l'utilisateur est bien administrateur.
+    if($permission > 1){
+        $pdoDe = $conn->prepare('DELETE FROM utilisateurs WHERE id = ? LIMIT 1');
+        $pdoDe->execute([
+            nettoyage($_GET['id'])
+        ]);
+    }
+}
+
 header("Location: ../gestion_user.php");
 ?>
